requirements.txt
# Python 3.14 required
fastapi>=0.115,<0.129
uvicorn[standard]>=0.32
tortoise-orm>=0.25.2
asyncpg>=0.30.0
aiomysql>=0.2.0
celery[redis]>=5.6
redis>=5.0
bcrypt>=4.3.0
pyjwt[crypto]>=2.8
python-dotenv>=1.1.1
uvloop>=0.21.0
pydantic[email]>=2.10,<3
click>=8.1.0
aerich[toml]>=0.7.0
